1.代码结构大调整 2.干掉jdbi

+++ gms-server/src/main/java/org/gms/dao/entity/NotesDO.java
@@ -3,14 +3,13 @@ package org.gms.dao.entity;
 import com.mybatisflex.annotation.Id;
 import com.mybatisflex.annotation.KeyType;
 import com.mybatisflex.annotation.Table;
-import java.io.Serializable;
-import java.math.BigInteger;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
 import java.io.Serial;
+import java.io.Serializable;
 
 /**
  *  实体类。
@@ -37,7 +36,7 @@ public class NotesDO implements Serializable {
 
     private String message;
 
-    private BigInteger timestamp;
+    private Long timestamp;
 
     private Integer fame;
